+++ pkg/core/faultinjections/matcher_suite_test.go
@@ -1,0 +1,13 @@
+package faultinjections
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestMatcher(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "FaultInjection Matcher Suite")
+}